Beginning to work on UI
began construction of more polished UI with far more capabilities. Meant to actually be interacted with

+++ src/Graphics/GUI/ToggleBox.cpp
@@ -21,7 +21,7 @@ ToggleBox::ToggleBox(
 
 	outBox.setFillColor(backgroundColor);
 	outBox.setOutlineColor(outlineColor);
-	outBox.setOutlineThickness(outlineWidth);
+	outBox.setOutlineThickness((float)outlineWidth);
 	
 	selectionBox.setFillColor(backgroundColor);
 	selectionBox.setOutlineThickness(0);
